Fix spacefn bug: should replay only keypresses from queue
This fixes a bug, which can be reproduced as follow. Suppose you
intend to type a space followed by an exclamation mark. You (1) press
space, (2) press shift, (3) release space, (4) press 1, (5) release 1,
and (6) release shift. At step (3), spacefn understands that you
intended to type a space, so it emits the space and removes the shift
from the rollover buffer. The bug was that it emitted both a press and
release for the shift when removing it from the buffer. That is not
correct: it made the exclamation mark show up as a "1".